requirements.txt
python-telegram-bot==21.5
gspread==6.1.2
google-auth==2.34.0
